Merge branch 'feature/files-bytes-bigint' into 'main'
* change postgresql INTEGER to BIGINT in files table to allow files larger than 2 GB

See merge request https://gitlab.ci.csc.fi/sds-dev/sd-submit/metadata-submitter/-/merge_requests/1074

Approved-by: Rasko Leinonen <raskolei@csc.fi>
Merged by Rasko Leinonen <raskolei@csc.fi>

+++ metadata_backend/database/postgres/schema/create.sql
@@ -63,7 +63,7 @@ CREATE TABLE files (
 	submission_id VARCHAR(128) NOT NULL,
 	object_id VARCHAR(128),
 	path VARCHAR(1024) NOT NULL,
-	bytes INTEGER,
+	bytes BIGINT,
 	checksum_method VARCHAR(16),
 	unencrypted_checksum VARCHAR(128),
 	encrypted_checksum VARCHAR(128),
